Add support for Visual Studio with the new native plugins
