+++ subprojects/ide-native/src/integTest/groovy/org/gradle/ide/visualstudio/VisualStudioSingleProjectIntegrationTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013 the original author or authors.
+ * Copyright 2018 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,870 +13,208 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.gradle.ide.visualstudio
 
 import org.gradle.ide.visualstudio.fixtures.AbstractVisualStudioIntegrationSpec
-import org.gradle.ide.visualstudio.fixtures.MSBuildExecutor
-import org.gradle.integtests.fixtures.SourceFile
 import org.gradle.internal.os.OperatingSystem
-import org.gradle.nativeplatform.fixtures.RequiresInstalledToolChain
 import org.gradle.nativeplatform.fixtures.app.CppHelloWorldApp
-import org.gradle.nativeplatform.fixtures.app.ExeWithDiamondDependencyHelloWorldApp
-import org.gradle.nativeplatform.fixtures.app.ExeWithLibraryUsingLibraryHelloWorldApp
-import org.gradle.nativeplatform.fixtures.app.MixedLanguageHelloWorldApp
-import org.gradle.nativeplatform.fixtures.app.WindowsResourceHelloWorldApp
-import org.gradle.util.Requires
-import org.gradle.util.TestPrecondition
-import spock.lang.Issue
 
-import static org.gradle.nativeplatform.fixtures.ToolChainRequirement.VISUALCPP
 
 class VisualStudioSingleProjectIntegrationTest extends AbstractVisualStudioIntegrationSpec {
-    private final Set<String> projectConfigurations = ['win32Debug', 'win32Release', 'x64Debug', 'x64Release'] as Set
+    private final Set<String> projectConfigurations = ['debug', 'release'] as Set
 
     def app = new CppHelloWorldApp()
 
     def setup() {
-        settingsFile << """
-            rootProject.name = 'app'
-        """
         buildFile << """
-            apply plugin: 'cpp'
             apply plugin: 'visual-studio'
-            
-            model {
-                platforms {
-                    win32 {
-                        architecture "i386"
-                    }
-                    x64 {
-                        architecture "amd64"
-                    }
-                }
-                buildTypes {
-                    debug
-                    release
-                }
-                components {
-                    all {
-                        targetPlatform "win32"
-                        targetPlatform "x64"
-                    }
-                }
-            }
         """
-    }
-
-    @Issue("https://github.com/gradle/gradle/issues/790")
-    def "creating visual studio multiple time gives the same result"() {
-        given:
-        app.writeSources(file("src/main"))
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec)
-    }
-}
-"""
-        when:
-        run "visualStudio"
-        def filtersFileContent = filtersFile("mainExe.vcxproj.filters").file.text
-        def projectFileContent = projectFile("mainExe.vcxproj").projectFile.text
-        def solutionFileContent = solutionFile("app.sln").file.text
-
-        then:
-        executedAndNotSkipped getExecutableTasks("main")
-
-        when:
-        run "visualStudio"
-
-        then:
-        executedAndNotSkipped getExecutableTasks("main")
-
-        and:
-        filtersFile("mainExe.vcxproj.filters").file.text == filtersFileContent
-        projectFile("mainExe.vcxproj").projectFile.text == projectFileContent
-        solutionFile("app.sln").file.text == solutionFileContent
     }
 
     def "create visual studio solution for single executable"() {
         when:
         app.writeSources(file("src/main"))
+        settingsFile << """
+            rootProject.name = 'app'
+        """
         buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            binaries.all {
-                cppCompiler.define "TEST"
-                cppCompiler.define "foo", "bar"
+            apply plugin: 'cpp-application'
+            
+            application {
+                binaries.configureEach { binary ->
+                    binary.compileTask.get().macros["TEST"] = null
+                    binary.compileTask.get().macros["foo"] = "bar"
+                }
             }
-        }
-    }
-}
-"""
+        """
+
         and:
         run "visualStudio"
 
         then:
-        executedAndNotSkipped ":visualStudio"
-        executedAndNotSkipped getExecutableTasks("main")
+        executedAndNotSkipped ":visualStudio", ":appVisualStudioSolution"
+        executedAndNotSkipped getProjectTasks("app")
 
         and:
-        final projectFile = projectFile("mainExe.vcxproj")
+        final projectFile = projectFile("app.vcxproj")
         projectFile.assertHasComponentSources(app, "src/main")
         projectFile.projectConfigurations.keySet() == projectConfigurations
         projectFile.projectConfigurations.values().each {
             assert it.macros == "TEST;foo=bar"
             assert it.includePath == filePath("src/main/headers")
-            assert it.buildCommand.endsWith("gradle\" :installMain${it.name.capitalize()}Executable")
-            assert it.outputFile == OperatingSystem.current().getExecutableName("build/install/main/${it.outputDir}/lib/main")
+            assert it.buildCommand.endsWith("gradle\" :install${it.name.capitalize()}")
+            assert it.outputFile == OperatingSystem.current().getExecutableName("build/exe/main/${it.name.toLowerCase()}/app")
         }
 
         and:
         final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe")
+        mainSolution.assertHasProjects("app")
         mainSolution.assertReferencesProject(projectFile, projectConfigurations)
     }
 
-    @Requires(TestPrecondition.MSBUILD)
-    def "can build executable from visual studio"() {
-        useMsbuildTool()
-        def debugBinary = executable("build/exe/main/win32/debug/main")
-
-        given:
-        app.writeSources(file("src/main"))
-        buildFile << """
-            model {
-                components {
-                    main(NativeExecutableSpec)
-                }
-            }
-        """
-
-        and:
-        succeeds "visualStudio"
-
-        when:
-        debugBinary.assertDoesNotExist()
-        def resultDebug = msbuild
-            .withSolution(solutionFile("app.sln"))
-            .withConfiguration('win32Debug')
-            .succeeds()
-
-        then:
-        resultDebug.assertTasksExecuted(':compileMainWin32DebugExecutableMainCpp', ':linkMainWin32DebugExecutable', ':mainWin32DebugExecutable', ':installMainWin32DebugExecutable')
-        resultDebug.assertTasksNotSkipped(':compileMainWin32DebugExecutableMainCpp', ':linkMainWin32DebugExecutable', ':mainWin32DebugExecutable', ':installMainWin32DebugExecutable')
-        debugBinary.assertExists()
-        installation('build/install/main/win32/debug').assertInstalled()
-    }
-
-    @Requires(TestPrecondition.MSBUILD)
-    def "can build library from visual studio"() {
-        useMsbuildTool()
-        def debugBinaryLib = staticLibrary("build/libs/main/static/win32/debug/main")
-        def debugBinaryDll = sharedLibrary("build/libs/main/shared/win32/debug/main")
-
-        given:
-        app.library.writeSources(file("src/main"))
-        buildFile << """
-model {
-    components {
-        main(NativeLibrarySpec)
-    }
-}
-"""
-        and:
-        succeeds "visualStudio"
-
-        when:
-        debugBinaryLib.assertDoesNotExist()
-        debugBinaryDll.assertDoesNotExist()
-        def resultDebug = msbuild
-            .withSolution(solutionFile("app.sln"))
-            .withConfiguration('win32Debug')
-            .succeeds()
-
-        then:
-        resultDebug.assertTasksExecuted(':compileMainWin32DebugStaticLibraryMainCpp', ':createMainWin32DebugStaticLibrary', ':mainWin32DebugStaticLibrary', ':compileMainWin32DebugSharedLibraryMainCpp', ':linkMainWin32DebugSharedLibrary', ':mainWin32DebugSharedLibrary')
-        resultDebug.assertTasksNotSkipped(':compileMainWin32DebugStaticLibraryMainCpp', ':createMainWin32DebugStaticLibrary', ':mainWin32DebugStaticLibrary', ':compileMainWin32DebugSharedLibraryMainCpp', ':linkMainWin32DebugSharedLibrary', ':mainWin32DebugSharedLibrary')
-        debugBinaryLib.assertExists()
-        debugBinaryDll.assertExists()
-    }
-
-    @Requires(TestPrecondition.MSBUILD)
-    def "can detect build failure from visual studio"() {
-        useMsbuildTool()
-
-        given:
-        app.writeSources(file("src/main"))
-        file('src/main/cpp/broken.cpp') << 'Broken!'
-        buildFile << """
-            model {
-                components {
-                    main(NativeExecutableSpec)
-                }
-            }
-        """
-
-        and:
-        succeeds "visualStudio"
-
-        when:
-        def resultDebug = msbuild
-            .withSolution(solutionFile("app.sln"))
-            .withConfiguration('win32Debug')
-            .fails()
-
-        then:
-        resultDebug.assertOutputContains("broken.cpp(1): error C2143: syntax error: missing ';' before '!'")
-    }
-
-    @Requires(TestPrecondition.MSBUILD)
-    def "can clean from visual studio"() {
-        useMsbuildTool()
-        def debugBinary = executable('build/exe/main/win32/debug/main')
-
-        given:
-        app.writeSources(file("src/main"))
-        buildFile << """
-            model {
-                components {
-                    main(NativeExecutableSpec)
-                }
-            }
-        """
-
-        and:
-        succeeds "visualStudio"
-
-        when:
-        debugBinary.assertDoesNotExist()
-        msbuild
-            .withSolution(solutionFile("app.sln"))
-            .withConfiguration('win32Debug')
-            .succeeds()
-
-        then:
-        debugBinary.exec().out == app.englishOutput
-
-        when:
-        msbuild
-            .withSolution(solutionFile("app.sln"))
-            .withConfiguration('win32Debug')
-            .succeeds(MSBuildExecutor.MSBuildAction.CLEAN)
-
-        then:
-        file("build").assertDoesNotExist()
-    }
-
-    def "create visual studio solution for library"() {
+    def "create visual studio solution for single shared library"() {
         when:
         app.library.writeSources(file("src/main"))
+        settingsFile << """
+            rootProject.name = 'lib'
+        """
         buildFile << """
-model {
-    components {
-        main(NativeLibrarySpec)
-    }
-}
-"""
+            apply plugin: 'cpp-library'
+            
+            library {
+                binaries.configureEach { binary ->
+                    binary.compileTask.get().macros["TEST"] = null
+                    binary.compileTask.get().macros["foo"] = "bar"
+                }
+            }
+        """
+
         and:
         run "visualStudio"
 
         then:
-        executedAndNotSkipped ":visualStudio"
-        executedAndNotSkipped getLibraryTasks("main")
+        executedAndNotSkipped ":visualStudio", ":libVisualStudioSolution"
+        executedAndNotSkipped getProjectTasks("libDll")
 
         and:
-        def dllProjectFile = projectFile("mainDll.vcxproj")
-        dllProjectFile.assertHasComponentSources(app.library, "src/main")
-        dllProjectFile.projectConfigurations.keySet() == projectConfigurations
-        dllProjectFile.projectConfigurations.values().each {
-            assert it.includePath == filePath("src/main/headers")
-            assert it.buildCommand.endsWith("gradle\" :main${it.name.capitalize()}SharedLibrary")
-            assert it.outputFile == OperatingSystem.current().getSharedLibraryName("build/libs/main/shared/${it.outputDir}/main")
+        final projectFile = projectFile("libDll.vcxproj")
+        projectFile.assertHasComponentSources(app.library, "src/main")
+        projectFile.projectConfigurations.keySet() == projectConfigurations
+        projectFile.projectConfigurations.values().each {
+            assert it.macros == "TEST;foo=bar"
+            assert it.includePath == filePath("src/main/public", "src/main/headers")
+            assert it.buildCommand.endsWith("gradle\" :link${it.name.capitalize()}")
+            assert it.outputFile == OperatingSystem.current().getSharedLibraryName("build/lib/main/${it.name.toLowerCase()}/${stripped(it.name)}lib")
         }
 
         and:
-        def libProjectFile = projectFile("mainLib.vcxproj")
+        final mainSolution = solutionFile("lib.sln")
+        mainSolution.assertHasProjects("libDll")
+        mainSolution.assertReferencesProject(projectFile, projectConfigurations)
+    }
+
+    def "create visual studio solution for single static library"() {
+        when:
+        app.library.writeSources(file("src/main"))
+        settingsFile << """
+            rootProject.name = 'lib'
+        """
+        buildFile << """
+            apply plugin: 'cpp-library'
+            
+            library {
+                linkage = [Linkage.STATIC]
+                binaries.configureEach { binary ->
+                    binary.compileTask.get().macros["TEST"] = null
+                    binary.compileTask.get().macros["foo"] = "bar"
+                }
+            }
+        """
+
+        and:
+        run "visualStudio"
+
+        then:
+        executedAndNotSkipped ":visualStudio", ":libVisualStudioSolution"
+        executedAndNotSkipped getProjectTasks("libLib")
+
+        and:
+        final projectFile = projectFile("libLib.vcxproj")
+        projectFile.assertHasComponentSources(app.library, "src/main")
+        projectFile.projectConfigurations.keySet() == projectConfigurations
+        projectFile.projectConfigurations.values().each {
+            assert it.macros == "TEST;foo=bar"
+            assert it.includePath == filePath("src/main/public", "src/main/headers")
+            assert it.buildCommand.endsWith("gradle\" :create${it.name.capitalize()}")
+            assert it.outputFile == OperatingSystem.current().getStaticLibraryName("build/lib/main/${it.name.toLowerCase()}/lib")
+        }
+
+        and:
+        final mainSolution = solutionFile("lib.sln")
+        mainSolution.assertHasProjects("libLib")
+        mainSolution.assertReferencesProject(projectFile, projectConfigurations)
+    }
+
+    def "create visual studio solution for single library with both static and shared linkages"() {
+        when:
+        app.library.writeSources(file("src/main"))
+        settingsFile << """
+            rootProject.name = 'lib'
+        """
+        buildFile << """
+            apply plugin: 'cpp-library'
+            
+            library {
+                linkage = [Linkage.STATIC, Linkage.SHARED]
+                binaries.configureEach { binary ->
+                    binary.compileTask.get().macros["TEST"] = null
+                    binary.compileTask.get().macros["foo"] = "bar"
+                }
+            }
+        """
+
+        and:
+        run "visualStudio"
+
+        then:
+        executedAndNotSkipped ":visualStudio", ":libVisualStudioSolution"
+        executedAndNotSkipped getProjectTasks("libLib")
+        executedAndNotSkipped getProjectTasks("libDll")
+
+        and:
+        final libProjectFile = projectFile("libLib.vcxproj")
         libProjectFile.assertHasComponentSources(app.library, "src/main")
         libProjectFile.projectConfigurations.keySet() == projectConfigurations
         libProjectFile.projectConfigurations.values().each {
-            assert it.includePath == filePath("src/main/headers")
-            assert it.buildCommand.endsWith("gradle\" :main${it.name.capitalize()}StaticLibrary")
-            assert it.outputFile == OperatingSystem.current().getStaticLibraryName("build/libs/main/static/${it.outputDir}/main")
+            assert it.macros == "TEST;foo=bar"
+            assert it.includePath == filePath("src/main/public", "src/main/headers")
+            assert it.buildCommand.endsWith("gradle\" :create${it.name.capitalize()}Static")
+            assert it.outputFile == OperatingSystem.current().getStaticLibraryName("build/lib/main/${it.name.toLowerCase()}/static/lib")
         }
 
         and:
-        def mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainDll", "mainLib")
+        final dllProjectFile = projectFile("libDll.vcxproj")
+        dllProjectFile.assertHasComponentSources(app.library, "src/main")
+        dllProjectFile.projectConfigurations.keySet() == projectConfigurations
+        dllProjectFile.projectConfigurations.values().each {
+            assert it.macros == "TEST;foo=bar"
+            assert it.includePath == filePath("src/main/public", "src/main/headers")
+            assert it.buildCommand.endsWith("gradle\" :link${it.name.capitalize()}Shared")
+            assert it.outputFile == OperatingSystem.current().getSharedLibraryName("build/lib/main/${it.name.toLowerCase()}/shared/${stripped(it.name)}lib")
+        }
+
+        and:
+        final mainSolution = solutionFile("lib.sln")
+        mainSolution.assertHasProjects("libLib", "libDll")
         mainSolution.assertReferencesProject(dllProjectFile, projectConfigurations)
     }
 
-    def "create visual studio solution for build with an executable and library"() {
-        when:
-        app.executable.writeSources(file("src/main"))
-        app.library.writeSources(file("src/hello"))
-        buildFile << """
-model {
-    components {
-        hello(NativeLibrarySpec)
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.lib library: 'hello', linkage: 'static'
-            }
-        }
-    }
-}
-"""
-        and:
-        run "visualStudio"
-
-        then:
-        final exeProject = projectFile("mainExe.vcxproj")
-        exeProject.assertHasComponentSources(app.executable, "src/main")
-        exeProject.projectConfigurations.keySet() == projectConfigurations
-        exeProject.projectConfigurations['win32Debug'].includePath == filePath("src/main/headers", "src/hello/headers")
-
-        and:
-        final libProject = projectFile("helloLib.vcxproj")
-        libProject.assertHasComponentSources(app.library, "src/hello")
-        libProject.projectConfigurations.keySet() == projectConfigurations
-        libProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final dllProject = projectFile("helloDll.vcxproj")
-        dllProject.assertHasComponentSources(app.library, "src/hello")
-        dllProject.projectConfigurations.keySet() == projectConfigurations
-        dllProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib")
-        mainSolution.assertReferencesProject(exeProject, projectConfigurations)
-        mainSolution.assertReferencesProject(libProject, projectConfigurations)
+    private String[] getProjectTasks(String exeName) {
+        return [":${exeName}VisualStudioProject", ":${exeName}VisualStudioFilters"]
     }
 
-    def "create visual studio project for executable that targets multiple platforms with the same architecture"() {
-        when:
-        app.writeSources(file("src/main"))
-        buildFile << """
-model {
-    platforms {
-        otherWin32 {
-            architecture "i386"
-        }
-    }
-    components {
-        main(NativeExecutableSpec) {
-            targetPlatform "otherWin32"
-        }
-    }
-}
-"""
-        and:
-        run "visualStudio"
-
-        then:
-        final mainProjectFile = projectFile("mainExe.vcxproj")
-        mainProjectFile.projectConfigurations.keySet() == ['win32Debug', 'otherWin32Debug', 'win32Release', 'otherWin32Release', 'x64Debug', 'x64Release'] as Set
-    }
-
-    def "create visual studio solution for executable that has diamond dependency"() {
-        def testApp = new ExeWithDiamondDependencyHelloWorldApp()
-        testApp.writeSources(file("src/main"), file("src/hello"), file("src/greetings"))
-
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.lib library: "hello"
-                cpp.lib library: "greetings", linkage: "static"
-            }
-        }
-        hello(NativeLibrarySpec) {
-            sources {
-                cpp.lib library: "greetings", linkage: "static"
-            }
-        }
-        greetings(NativeLibrarySpec)
-    }
-}
-"""
-
-        when:
-        succeeds "visualStudio"
-
-        then:
-        final exeProject = projectFile("mainExe.vcxproj")
-        final helloDllProject = projectFile("helloDll.vcxproj")
-        final helloLibProject = projectFile("helloDll.vcxproj")
-        final greetDllProject = projectFile("greetingsLib.vcxproj")
-        final greetLibProject = projectFile("greetingsLib.vcxproj")
-        final mainSolution = solutionFile("app.sln")
-
-        and:
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib", "greetingsDll", "greetingsLib")
-        mainSolution.assertReferencesProject(exeProject, projectConfigurations)
-        mainSolution.assertReferencesProject(helloDllProject, projectConfigurations)
-        mainSolution.assertReferencesProject(helloLibProject, projectConfigurations)
-        mainSolution.assertReferencesProject(greetDllProject, projectConfigurations)
-        mainSolution.assertReferencesProject(greetLibProject, projectConfigurations)
-
-        and:
-        exeProject.assertHasComponentSources(testApp.executable, "src/main")
-        exeProject.projectConfigurations.keySet() == projectConfigurations
-        exeProject.projectConfigurations['win32Debug'].includePath == filePath("src/main/headers", "src/hello/headers", "src/greetings/headers")
-    }
-
-    def "generate visual studio solution for executable with mixed sources"() {
-        given:
-        def testApp = new MixedLanguageHelloWorldApp(toolChain)
-        testApp.writeSources(file("src/main"))
-
-        and:
-        buildFile << """
-apply plugin: 'c'
-apply plugin: 'assembler'
-model {
-    components {
-        main(NativeExecutableSpec)
-    }
-}
-"""
-
-        when:
-        run "visualStudio"
-
-        then:
-        final projectFile = projectFile("mainExe.vcxproj")
-        projectFile.assertHasComponentSources(testApp, "src/main")
-        projectFile.projectConfigurations.keySet() == projectConfigurations
-        with (projectFile.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/main/headers")
-        }
-
-        and:
-        solutionFile("app.sln").assertHasProjects("mainExe")
-    }
-
-    @RequiresInstalledToolChain(VISUALCPP)
-    def "generate visual studio solution for executable with windows resource files"() {
-        given:
-        def resourceApp = new WindowsResourceHelloWorldApp()
-        resourceApp.writeSources(file("src/main"))
-
-        and:
-        buildFile << """
-apply plugin: 'windows-resources'
-model {
-    components {
-        main(NativeExecutableSpec)
-    }
-    binaries {
-        all {
-            rcCompiler.define "TEST"
-            rcCompiler.define "foo", "bar"
-        }
-    }
-}
-"""
-
-        when:
-        run "visualStudio"
-
-        then:
-        final projectFile = projectFile("mainExe.vcxproj")
-        final List<SourceFile> resources = resourceApp.resourceSources
-        final List<SourceFile> sources = resourceApp.sourceFiles - resources
-        assert projectFile.headerFiles == resourceApp.headerFiles*.withPath("src/main").sort()
-        assert projectFile.sourceFiles == ['build.gradle'] + sources*.withPath("src/main").sort()
-        assert projectFile.resourceFiles == resources*.withPath("src/main").sort()
-
-        projectFile.projectConfigurations.keySet() == projectConfigurations
-        with (projectFile.projectConfigurations['win32Debug']) {
-            macros == "TEST;foo=bar"
-            includePath == filePath("src/main/headers")
-        }
-
-        and:
-        solutionFile("app.sln").assertHasProjects("mainExe")
-    }
-
-    def "builds solution for component with no source"() {
-        given:
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec)
-    }
-}
-"""
-
-        when:
-        run "visualStudio"
-
-        then:
-        final projectFile = projectFile("mainExe.vcxproj")
-        projectFile.sourceFiles == ['build.gradle']
-        projectFile.headerFiles == []
-        projectFile.projectConfigurations.keySet() == projectConfigurations
-        with (projectFile.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/main/headers")
-        }
-
-        and:
-        solutionFile("app.sln").assertHasProjects("mainExe")
-    }
-
-    def "visual studio project includes headers co-located with sources"() {
-        when:
-        // Write headers so they sit with sources
-        app.files.each {
-            it.writeToFile(file("src/main/cpp/${it.name}"))
-        }
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.source.include "**/*.cpp"
-            }
-        }
-    }
-}
-"""
-        and:
-        run "mainVisualStudio"
-
-        then:
-        executedAndNotSkipped getExecutableTasks("main")
-
-        and:
-        final projectFile = projectFile("mainExe.vcxproj")
-        assert projectFile.sourceFiles == ['build.gradle'] + app.sourceFiles.collect({"src/main/cpp/${it.name}"}).sort()
-        assert projectFile.headerFiles == app.headerFiles.collect({"src/main/cpp/${it.name}"}).sort()
-    }
-
-    def "visual studio solution with header-only library"() {
-        given:
-        def app = new CppHelloWorldApp()
-        app.executable.writeSources(file("src/main"))
-
-        app.library.headerFiles*.writeToDir(file("src/helloApi"))
-        app.library.sourceFiles*.writeToDir(file("src/hello"))
-
-        and:
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.lib library: 'helloApi', linkage: 'api' // TODO:DAZ This should not be needed
-                cpp.lib library: 'hello'
-            }
-        }
-        helloApi(NativeLibrarySpec)
-        hello(NativeLibrarySpec) {
-            sources {
-                cpp.lib library: 'helloApi', linkage: 'api'
-            }
-        }
-    }
-}
-"""
-
-        when:
-        succeeds "visualStudio"
-
-        then:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib", "helloApiDll", "helloApiLib")
-
-        and:
-        final mainExeProject = projectFile("mainExe.vcxproj")
-        with (mainExeProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/main/headers", "src/helloApi/headers", "src/hello/headers")
-        }
-
-        and:
-        final helloDllProject = projectFile("helloDll.vcxproj")
-        with (helloDllProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/hello/headers", "src/helloApi/headers")
-        }
-
-        and:
-        final helloLibProject = projectFile("helloLib.vcxproj")
-        with (helloLibProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/hello/headers", "src/helloApi/headers")
-        }
-
-        and:
-        final helloApiDllProject = projectFile("helloApiDll.vcxproj")
-        with (helloApiDllProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/helloApi/headers")
-        }
-
-        and:
-        final helloApiLibProject = projectFile("helloApiLib.vcxproj")
-        with (helloApiLibProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/helloApi/headers")
-        }
-    }
-
-    def "create visual studio solution for executable with variant conditional sources"() {
-        when:
-        app.writeSources(file("src/win32"))
-        app.alternate.writeSources(file("src/x64"))
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            binaries.all { binary ->
-                def platformName = binary.targetPlatform.name
-                sources {
-                    platformSources(CppSourceSet) {
-                        source.srcDir "src/\$platformName/cpp"
-                        exportedHeaders.srcDir "src/\$platformName/headers"
-                    }
-                }
-            }
-        }
-    }
-}
-
-"""
-        and:
-        run "visualStudio"
-
-        then:
-        final projectFile = projectFile("mainExe.vcxproj")
-        projectFile.assertHasComponentSources(app, "src/win32", app.alternate, "src/x64")
-        projectFile.projectConfigurations.keySet() == projectConfigurations
-
-        and:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe")
-        mainSolution.assertReferencesProject(projectFile, projectConfigurations)
-    }
-
-    def "visual studio solution with pre-built library"() {
-        given:
-        app.writeSources(file("src/main"))
-        buildFile << """
-model {
-    repositories {
-        libs(PrebuiltLibraries) {
-            test {
-                headers.srcDir "libs/test/include"
-            }
-        }
-    }
-    components {
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.lib library: 'test', linkage: 'api'
-            }
-        }
-    }
-}
-"""
-
-        when:
-        run "visualStudio"
-
-        then:
-        executedAndNotSkipped getExecutableTasks("main")
-        and:
-
-        then:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe")
-
-        and:
-        final mainExeProject = projectFile("mainExe.vcxproj")
-        with (mainExeProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/main/headers", "libs/test/include")
-        }
-    }
-
-    def "visual studio solution for executable that depends on library using precompiled header"() {
-        when:
-        app = new CppHelloWorldApp()
-        app.executable.writeSources(file("src/main"))
-        app.library.writeSources(file("src/hello"))
-        buildFile << """
-            model {
-                components {
-                    hello(NativeLibrarySpec) {
-                        sources {
-                            cpp.preCompiledHeader "pch.h"
-                        }
-                    }
-                    main(NativeExecutableSpec) {
-                        sources {
-                            cpp.lib library: 'hello'
-                        }
-                    }
-                }
-            }
-        """
-        and:
-        run "visualStudio"
-
-        then:
-        final exeProject = projectFile("mainExe.vcxproj")
-        exeProject.assertHasComponentSources(app.executable, "src/main")
-        exeProject.projectConfigurations.keySet() == projectConfigurations
-        exeProject.projectConfigurations['win32Debug'].includePath == filePath("src/main/headers", "src/hello/headers")
-
-        and:
-        final dllProject = projectFile("helloDll.vcxproj")
-        dllProject.assertHasComponentSources(app.library, "src/hello")
-        dllProject.projectConfigurations.keySet() == projectConfigurations
-        dllProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final libProject = projectFile("helloLib.vcxproj")
-        libProject.assertHasComponentSources(app.library, "src/hello")
-        libProject.projectConfigurations.keySet() == projectConfigurations
-        libProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib")
-        mainSolution.assertReferencesProject(exeProject, projectConfigurations)
-        mainSolution.assertReferencesProject(dllProject, projectConfigurations)
-    }
-
-    def "visual studio solution for component graph with library dependency cycle"() {
-        given:
-        def app = new ExeWithLibraryUsingLibraryHelloWorldApp()
-        app.executable.writeSources(file("src/main"))
-        app.library.writeSources(file("src/hello"))
-        app.greetingsHeader.writeToDir(file("src/hello"))
-        app.greetingsSources*.writeToDir(file("src/greetings"))
-
-        and:
-        buildFile << """
-model {
-    components {
-        main(NativeExecutableSpec) {
-            sources {
-                cpp.lib library: 'hello'
-            }
-        }
-        hello(NativeLibrarySpec) {
-            sources {
-                cpp.lib library: 'greetings', linkage: 'static'
-            }
-        }
-        greetings(NativeLibrarySpec) {
-            sources {
-                cpp.lib library: 'hello', linkage: 'api'
-            }
-        }
-    }
-}
-"""
-
-        when:
-        succeeds "visualStudio"
-
-        then:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib", "greetingsLib", "greetingsDll")
-
-        and:
-        final mainExeProject = projectFile("mainExe.vcxproj")
-        with (mainExeProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/main/headers", "src/hello/headers")
-        }
-
-        and:
-        final helloDllProject = projectFile("helloDll.vcxproj")
-        with (helloDllProject.projectConfigurations['win32Debug']) {
-            includePath == filePath( "src/hello/headers", "src/greetings/headers")
-        }
-
-        and:
-        final helloLibProject = projectFile("helloLib.vcxproj")
-        with (helloLibProject.projectConfigurations['win32Debug']) {
-            includePath == filePath( "src/hello/headers", "src/greetings/headers")
-        }
-
-        and:
-        final greetingsDllProject = projectFile("greetingsDll.vcxproj")
-        with (greetingsDllProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/greetings/headers", "src/hello/headers")
-        }
-
-        and:
-        final greetingsLibProject = projectFile("greetingsLib.vcxproj")
-        with (greetingsLibProject.projectConfigurations['win32Debug']) {
-            includePath == filePath("src/greetings/headers", "src/hello/headers")
-        }
-    }
-
-    def "create visual studio solution where referenced projects have different configurations"() {
-        when:
-        app.executable.writeSources(file("src/main"))
-        app.library.writeSources(file("src/hello"))
-        buildFile << """
-model {
-    components {
-        hello(NativeLibrarySpec)
-        main(NativeExecutableSpec) {
-            targetBuildTypes "release"
-            sources {
-                cpp.lib library: 'hello'
-            }
-        }
-    }
-}
-"""
-        and:
-        run "visualStudio"
-
-        then:
-        final exeProject = projectFile("mainExe.vcxproj")
-        exeProject.assertHasComponentSources(app.executable, "src/main")
-        exeProject.projectConfigurations.keySet() == ['win32', 'x64'] as Set
-        exeProject.projectConfigurations['win32'].includePath == filePath("src/main/headers", "src/hello/headers")
-
-        and:
-        final dllProject = projectFile("helloDll.vcxproj")
-        dllProject.assertHasComponentSources(app.library, "src/hello")
-        dllProject.projectConfigurations.keySet() == projectConfigurations
-        dllProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final libProject = projectFile("helloLib.vcxproj")
-        libProject.assertHasComponentSources(app.library, "src/hello")
-        libProject.projectConfigurations.keySet() == projectConfigurations
-        libProject.projectConfigurations['win32Debug'].includePath == filePath("src/hello/headers")
-
-        and:
-        final mainSolution = solutionFile("app.sln")
-        mainSolution.assertHasProjects("mainExe", "helloDll", "helloLib")
-        mainSolution.assertReferencesProject(exeProject, ['win32', 'x64'])
-        mainSolution.assertReferencesProject(dllProject, ['win32Debug', 'x64Debug', 'win32Release', 'x64Release'])
-        mainSolution.assertReferencesProject(libProject, ['win32Debug', 'x64Debug', 'win32Release', 'x64Release'])
-    }
-
-    private String[] getLibraryTasks(String libraryName) {
-        return getStaticLibraryTasks(libraryName) + getSharedLibraryTasks(libraryName)
-    }
-
-    private String[] getStaticLibraryTasks(String libraryName) {
-        return [":${libraryName}LibVisualStudioProject", ":${libraryName}LibVisualStudioFilters"]
-    }
-
-    private String[] getSharedLibraryTasks(String libraryName) {
-        return [":${libraryName}DllVisualStudioProject", ":${libraryName}DllVisualStudioFilters"]
-    }
-
-    private String[] getExecutableTasks(String exeName) {
-        return [":${exeName}ExeVisualStudioProject", ":${exeName}ExeVisualStudioFilters"]
+    private String stripped(String configurationName) {
+        return configurationName == "release" ? "stripped/" : ""
     }
 }
